# backend/src/utils/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NGIPS Phishing Shield API",
    "description": "Real-time phishing URL detection API for identifying malicious websites",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "tags": [
    {
      "name": "Analysis",
      "description": "URL phishing detection endpoints"
    },
    {
      "name": "Authentication",
      "description": "User authentication endpoints"
    },
    {
      "name": "Feedback",
      "description": "User feedback endpoints"
    },
    {
      "name": "Statistics",
      "description": "Aggregate statistics endpoints"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "description": "Returns the health status of the service",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "healthy"
                    },
                    "service": {
                      "type": "string",
                      "example": "ngips-phishing-shield"
                    },
                    "version": {
                      "type": "string",
                      "example": "1.0.0"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/analyze": {
      "post": {
        "tags": ["Analysis"],
        "summary": "Analyze URL for phishing",
        "description": "Analyzes a URL and returns phishing detection results",
        "operationId": "analyzeUrl",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url"],
                "properties": {
                  "url": {
                    "type": "string",
                    "description": "URL to analyze",
                    "example": "https://example.com/login"
                  },
                  "extensionVersion": {
                    "type": "string",
                    "description": "Browser extension version",
                    "example": "1.0.0"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Analysis successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/scans": {
      "get": {
        "tags": ["Analysis"],
        "summary": "Get scan history",
        "description": "Retrieves paginated scan history",
        "operationId": "getScans",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 100
            },
            "description": "Items per page"
          },
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["block", "allow"]
            },
            "description": "Filter by action"
          }
        ],
        "responses": {
          "200": {
            "description": "Scan history retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanHistory"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "description": "Authenticates user and returns access token",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "secret123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Refresh access token",
        "description": "Refreshes expired access token using refresh token",
        "operationId": "refreshToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "example": "abc123..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed"
          }
        }
      }
    },
    "/v1/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current user",
        "description": "Returns the currently authenticated user",
        "operationId": "getCurrentUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User info retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
    "/v1/feedback": {
      "post": {
        "tags": ["Feedback"],
        "summary": "Submit feedback",
        "description": "Submit feedback on a scan result",
        "operationId": "submitFeedback",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["scanId"],
                "properties": {
                  "scanId": {
                    "type": "string",
                    "example": "scan_abc123"
                  },
                  "isFalsePositive": {
                    "type": "boolean",
                    "example": true
                  },
                  "userComment": {
                    "type": "string",
                    "example": "This is a legitimate site"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Feedback submitted"
          },
          "404": {
            "description": "Scan not found"
          }
        }
      }
    },
    "/v1/stats": {
      "get": {
        "tags": ["Statistics"],
        "summary": "Get statistics",
        "description": "Returns aggregate statistics",
        "operationId": "getStats",
        "responses": {
          "200": {
            "description": "Statistics retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stats"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "scan_abc123"
          },
          "url": {
            "type": "string",
            "example": "https://example.com"
          },
          "action": {
            "type": "string",
            "enum": ["block", "allow"],
            "example": "block"
          },
          "confidence": {
            "type": "number",
            "example": 0.85
          },
          "threatLevel": {
            "type": "string",
            "enum": ["low", "medium", "high", "critical"],
            "example": "high"
          },
          "reasons": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Suspicious top-level domain"]
          },
          "processingTime": {
            "type": "number",
            "example": 15
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ScanHistory": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnalysisResult"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "number"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["user", "admin"]
          }
        }
      },
      "Stats": {
        "type": "object",
        "properties": {
          "totalScans": {
            "type": "number"
          },
          "blockedCount": {
            "type": "number"
          },
          "allowedCount": {
            "type": "number"
          },
          "blockRate": {
            "type": "number"
          },
          "threatLevelDistribution": {
            "type": "object",
            "properties": {
              "low": { "type": "number" },
              "medium": { "type": "number" },
              "high": { "type": "number" },
              "critical": { "type": "number" }
            }
          },
          "avgConfidence": {
            "type": "number"
          },
          "recentScansLast24h": {
            "type": "number"
          },
          "totalFeedback": {
            "type": "number"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}